requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
bidict==0.21.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
dateparser==1.1.0
Flask==2.0.2
Flask-SocketIO==5.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
multidict==5.2.0
python-binance==1.0.12
python-dateutil==2.8.2
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.3
PyYAML==5.4.1
regex==2021.9.30
requests==2.26.0
simplejson==3.17.3
six==1.16.0
tradingview-ta==3.2.7
typing-extensions==3.10.0.2
tzlocal==3.0
ujson==4.2.0
urllib3==1.26.7
websockets==10.0
Werkzeug==2.0.2
yarl==1.7.0
